Move log files to logs dir and also log query ID
